SUMMARY: Show Dropdown OnFocus, Hide on Blur or Click
A default ContactsList should be shown when a user clicks in the
SearchBar input. After a user clicks on a contact in the contacts list
or clicks on the Search button, the action associated with the
Contact or the Search button should fire, and then the ContactsList
should disappear.

This PR changes the toggling of the display of the Contacts list so
that it relies more on the render method than on CSS. It introduces
a toggle mask that creates a clickable area behind the displayed
SearchBar and ContactList Dropdown to toggle the hiding of the
ContactList Dropdown. It reverts the show and hide methods to a single
toggle method.